pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pradel-jax"
version = "2.0.0-alpha"
description = "Flexible and robust capture-recapture analysis using JAX"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ava Britton", email = "ava.britton@example.com"},
    {name = "Christopher Chizinski", email = "cchizinski2@unl.edu"},
]
keywords = [
    "capture-recapture", 
    "survival analysis", 
    "pradel model", 
    "wildlife ecology",
    "jax",
    "optimization"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
]
dependencies = [
    "jax>=0.4.20",
    "jaxlib>=0.4.20", 
    "jaxopt>=0.8.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "patsy>=0.5.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.64.0",
    "optax>=0.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-benchmark>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
validation = [
    "rpy2>=3.5.0",  # For R integration and RMark validation
]
gpu = [
    "jax[cuda]>=0.4.20",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/user/pradel-jax"
Documentation = "https://pradel-jax.readthedocs.io"
Repository = "https://github.com/user/pradel-jax"
"Bug Tracker" = "https://github.com/user/pradel-jax/issues"

[project.scripts]
pradel-jax = "pradel_jax.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pradel_jax*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "--cov=pradel_jax",
    "--cov-report=term-missing",
    "--cov-report=html",
]